+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 4.4.0
+* `image-build`
+  * New option `image-cache-ignore` / docker image label `tci.image-cache.ignore`
+    * If this option/label is set on an image it will not be saved when using the cache save action
+    * Can be used in situations where the built image can't be deleted but it should not be saved in the cache
+* Updated dependencies
+
 # 4.3.0
 * `selenium`
   * Selenium 4.47+ is now required
@@ -10,7 +17,7 @@
 * `oidc-server-mock`
   * Updated to `1.3`
   * Switched to Java's built-in HttpClient
-  * Extracted API into dedicated class
+  * Extracted API into dedicated classes
   * Added support for proper serialization (using Jackson) with dedicated models and builder
 * `image-build`
   * Add shortcut methods to `BuildImage`